Extract the formatter class from TextUi
